+++ src/open_subsonic/common/id3.rs
@@ -34,8 +34,7 @@ pub struct BasicAlbumId3 {
     pub id: Uuid,
     pub name: String,
     pub song_count: i64,
-    #[derivative(PartialEq = "ignore")]
-    pub duration: f32,
+    pub duration: i64,
     #[serde(with = "crate::utils::time::iso8601_datetime")]
     pub created: OffsetDateTime,
 }
@@ -57,8 +56,7 @@ pub struct AlbumId3 {
 pub struct BasicSongId3 {
     pub id: Uuid,
     pub title: String,
-    #[derivative(PartialEq = "ignore")]
-    pub duration: f32,
+    pub duration: i32,
     pub size: i64,
     #[serde(with = "crate::utils::time::iso8601_datetime")]
     pub created: OffsetDateTime,
