+++ src/utils/song/parse/information.rs
@@ -33,7 +33,7 @@ impl SongInformation {
 
                 let flac_property = flac_file.properties();
                 let song_property = SongProperty {
-                    duration: flac_property.duration().as_secs_f32(),
+                    duration: flac_property.duration().as_secs_f32() as u32,
                     bitrate: flac_property.audio_bitrate(),
                     sample_rate: flac_property.sample_rate(),
                     channel_count: flac_property.channels(),
@@ -52,7 +52,7 @@ impl SongInformation {
 
                 let mp3_property = mp3_file.properties();
                 let song_property = SongProperty {
-                    duration: mp3_property.duration().as_secs_f32(),
+                    duration: mp3_property.duration().as_secs_f32() as u32,
                     bitrate: mp3_property.audio_bitrate(),
                     sample_rate: mp3_property.sample_rate(),
                     channel_count: mp3_property.channels(),
@@ -109,7 +109,7 @@ impl SongInformation {
                 .map(|language| language.to_639_3())
                 .collect_vec(),
             // Song property
-            duration: self.property.duration,
+            duration: self.property.duration as i32,
             bitrate: self.property.bitrate as i32,
             sample_rate: self.property.sample_rate as i32,
             channel_count: self.property.channel_count as i16,
